Оптимизация + editOrder

+++ app/src/main/java/com/bmvl/lk/data/models/Orders.java
@@ -4,7 +4,7 @@ public class Orders {
     private int id;
     private String id1c;
     private int user_id;
-    private int type_id;
+    private byte type_id;
     private int status_id;
     private String number;
     private String ecp;
@@ -17,7 +17,7 @@ public class Orders {
     private int send_to_1c;
     private String act_of_selection;
 
-    public Orders(int id, int user_id, int type_id, int status_id, String date) {
+    public Orders(int id, int user_id, byte type_id, int status_id, String date) {
         this.id = id;
         this.user_id = user_id;
         this.type_id = type_id;
@@ -25,7 +25,7 @@ public class Orders {
         this.date = date;
     }
 
-    public Orders(int id, String id1c, int user_id, int type_id, int status_id, String number, String ecp, String date, String note, String ecp_note, double price, int deleted, int from1c, int send_to_1c, String act_of_selection) {
+    public Orders(int id, String id1c, int user_id, byte type_id, int status_id, String number, String ecp, String date, String note, String ecp_note, double price, int deleted, int from1c, int send_to_1c, String act_of_selection) {
         this.id = id;
         this.id1c = id1c;
         this.user_id = user_id;
@@ -55,7 +55,7 @@ public class Orders {
         return user_id;
     }
 
-    public int getType_id() {
+    public byte getType_id() {
         return type_id;
     }
